Create rock paper and scissors game functions
Created 4 functions for the game: game(), playRound(),playerChoice()
& machineChoice(). Each function takes care of an essential part
of the game. The game() function takes care of the turns and conditions
to win or lose a match. The playRound() function takes care of each
round and the conditions to win or lose a round. The playerChoice()
function gets the input from the user to play each round and the
machineChoice() function takes a random input for the machine to
play on each round.