Create a UI & a whole new pack of functionalities
Left behind the prompt/alert model and refactor the code to be more
interactive and functional. Create 3 buttons rock, paper & scissors
that register the player selection, initializing each match. Create a
series of functions to track the player's progress on a results
board.

The board contains, the round result message, the match result message
& the player's score, the machine's score and the draws while playing
each match. Once the first match finishes the result board reports the
amount of games won and played.

Create a Reset button that resets the whole game, it shows up when the
player initializes a match & it's gone when a new round begins or if
the player resets the game.

Create a Play Again button to start a new round it resets most of the
board only leaving unchanged the "Games won" and "Games played", the
button shows up when the player finishes each round and it's gone when
a new round begins or if the user resets the game.